requirements.txt
aiofiles==23.2.1
aiogram==2.13
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
Babel==2.15.0
black==24.4.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
frozenlist==1.4.1
idna==3.7
isort==5.13.2
magic-filter==1.0.12
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
pydantic==2.8.2
pydantic_core==2.20.1
pytz==2024.1
requests==2.32.3
schedule==1.2.2
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
